Moving around game tests in to seperate files in a folder to be a bit more organised
